Rename packages
resolves #13 and #18
