--- restli-tools-scala/src/main/scala-2.10/com/linkedin/restli/tools/scala/DocStringCompat.scala
@@ -1,9 +1,0 @@
-package com.linkedin.restli.tools.scala
-
-import scala.tools.nsc.doc.base._
-
-object DocStringCompat extends DocString {
-  override protected def linkToDocString: PartialFunction[LinkTo, String] = super.linkToDocString orElse {
-    case LinkToExternal(name, baseUrl) => s"""<a href="$baseUrl">$name</a>"""
-  }
-}
